backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.19.1
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.14
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
virtualenv==20.34.0
whitenoise==6.10.0
